refactor(BaseViews): authentication and permissions as separate modules
